refactor(grading): split LiveRunnerCallbackGradingSubstrate into substrate_live for runner-subset partition
Move LiveRunnerCallbackGradingSubstrate and its private gRPC helpers
(_GrpcDBReader, _GrpcKnowledgeSearch) from substrate.py to a new
substrate_live.py so the runner subset can exclude both substrate_client.py
and substrate_live.py without breaking the cross-import invariant that
locks the partition: substrate.py stays in the subset (runner/service.py
imports the GradingSubstrate Protocol + InProcessGradingSubstrate from it)
while the grader-side gRPC client tree is now free of any import from a
subset file, so both files ship base-wheel-only. Entry-point registration
points at the new module; three test files update their imports; the
runner-subset exclusion tuple and the pyproject `[tool.hatch.build.targets
.custom].exclude` list gain both paths in lockstep.
